Redesign ec tables

+++ app/models/shipping_rule.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, String, Boolean, DateTime, DECIMAL
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, DECIMAL
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 from app.database import Base
@@ -7,7 +7,7 @@ from app.database import Base
 class ShippingRule(Base):
     __tablename__ = "shipping_rules"
 
-    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     min_weight = Column(DECIMAL(10, 2))
     max_weight = Column(DECIMAL(10, 2))
     min_size = Column(DECIMAL(10, 2))
